requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
sqlalchemy==2.0.25
pydantic==2.5.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
authlib==1.2.0
aiogram==2.25.1
celery==5.3.4
pika==1.3.1